// src/types/saveData.d.ts
export type SaveData = {
  caseID?: string;
  caseNumber?: string;
  fillingDate?: string;
  description?: string;
  plaintiff?: string;
  defendant?: string;
  partyName?: string;
  partyType?: string;
  partyAddress1?: string;
  partyAddress2?: string;
  dispositionCode?: string;
  dispositionDate?: string;
  city?: string;
  state?: string;
  zip?: string;
};